Added generalised section_generators.
Up to now the plugin only had the [background_section]() generator; it
now also has typography_section and border_section. The next step is to
add an add_section_generator() method, which will allow you to create
custom pre-built sections.

The use case I have in mind for this is more elegant colour inheritance.
For example, say a theme's `primary_colour` and a `secondary_colour`
would be two settings. For any subsequent `colour` control, it'd be
great to have a set of radio buttons: `primary`, `secondary`, and
`custom`.

The first two options would apply that colour to the variable; the third
would display the colour control instead. Inside the Sass files you'd
create a mixin which returns the correct colour based on the radio
option selected.

This solution would be far better for usability, and much more in line
with how Sass works.