Setting Inheritance - first pass.
At first glance this looks like a fairly elegant solution to the
use-case I had in mind, and so far has worked perfectly in the few
situations I've thrown at it. I'll write up an explanation of what it
is, how it works, and how to configure it later (probably tomorrow), and
add that to the Readme file.